+++ feature/my/src/main/java/com/ilsangtech/ilsang/feature/my/navigation/MyTabNavigation.kt
@@ -12,10 +12,15 @@ import com.ilsangtech.ilsang.feature.my.MyTitleScreen
 import com.ilsangtech.ilsang.feature.my.UserProfileEditScreen
 import com.ilsangtech.ilsang.feature.my.screens.customer_center.CustomerCenterScreen
 import com.ilsangtech.ilsang.feature.my.screens.faq.FaqScreen
+import com.ilsangtech.ilsang.feature.my.SettingScreen
+import com.ilsangtech.ilsang.feature.my.TermsScreen
+import com.ilsangtech.ilsang.feature.my.WithdrawalScreen
+import com.ilsangtech.ilsang.feature.my.component.CustomerCenterScreen
 import com.ilsangtech.ilsang.feature.my.screens.mytab.MyTabScreen
 import com.ilsangtech.ilsang.feature.my.screens.setting.SettingScreen
 import com.ilsangtech.ilsang.feature.my.screens.terms.TermsScreen
 import com.ilsangtech.ilsang.feature.my.screens.withdrawal.WithdrawalScreen
+import com.ilsangtech.ilsang.feature.my.screens.profile_edit.UserProfileEditScreen
 import kotlinx.serialization.Serializable
 
 @Serializable
